requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
mysql-connector-python==9.3.0
python-dotenv==1.1.1
pydantic==2.11.7
twilio==8.9.1
